+++ src/Helpers/ImprovementHelper.php
@@ -26,7 +26,7 @@ class ImprovementHelper
         $ratingStrings = [
             'science' => '+%s%% platinum production',
             'keep' => '+%s%% max population',
-            'spires' => '+%s%% offensive wizard power, mana production, spell damage reduction',
+            'spires' => '+%s%% offensive wizard power, mana production, lightning damage reduction',
             'forges' => '+%s%% offensive power',
             'walls' => '+%s%% defensive power',
             'harbor' => '+%s%% food production, +%s%% boat production & protection',
@@ -42,7 +42,7 @@ class ImprovementHelper
         $helpStrings = [
             'science' => "Improvements to {$improvementName} increase your platinum production.<br><br>Max +20% base {$improvementName}.",
             'keep' => "Improvements to your {$improvementName} increase your maximum population.<br><br>Max +30% base {$improvementName}.",
-            'spires' => "Improvements to your {$improvementName} increase your offensive wizard power, mana production, and reduce damage from harmful spells.<br><br>Max +60% base {$improvementName}.",
+            'spires' => "Improvements to your {$improvementName} increase your offensive wizard power, mana production, and reduce damage from lightning bolt.<br><br>Max +60% base {$improvementName}.",
             'forges' => "Improvements to your {$improvementName} increase your offensive power.<br><br>Max +30% base {$improvementName}.",
             'walls' => "Improvements to your {$improvementName} increase your defensive power.<br><br>Max +30% base {$improvementName}.",
             'harbor' => "Improvements to your {$improvementName} improve your food production, boat production, and boat protection.<br><br>Max +60% base {$improvementName}.",